Add Pokémon controller, error handling, base specs
Add Pokémon controller with basic routing to a show action, taking
Pokémon name as a parameter. Return basic JSON to test functionality.

Add a route and basic error handling in Action Controller for missing
paths to prevent Runtime Errors when route is not found.

Add specs for the above, and minor changes to RSpec config for preferred
test output.